+++ src/main/java/com/minecraftonline/griefalert/commands/ApplierCommand.java
@@ -77,7 +77,7 @@ public final class ApplierCommand extends GeneralCommand {
     this.applyType = applyType;
     Lists.newArrayList(aliases).forEach(this::addAlias);
     setCommandElement(GenericArguments.flags()
-        .valueFlag(GenericArguments.string(CommandKeys.SINCE.get()), "s")
+        .valueFlag(GenericArguments.string(CommandKeys.AFTER.get()), "a")
         .valueFlag(GenericArguments.string(CommandKeys.BEFORE.get()), "b")
         .valueFlag(GenericArguments.string(CommandKeys.PLAYER.get()), "p")
         .valueFlag(GenericArguments.string(CommandKeys.PRISM_TARGET.get()), "t")
@@ -134,7 +134,7 @@ public final class ApplierCommand extends GeneralCommand {
 
 
         // Parse the 'since' with the given date format, or just do a year ago
-        Date since = args.<String>getOne(CommandKeys.SINCE.get()).flatMap(str -> {
+        Date since = args.<String>getOne(CommandKeys.AFTER.get()).flatMap(str -> {
           try {
             return Optional.of(DateUtil.parseAnyDate(str));
           } catch (IllegalArgumentException e) {