+++ src/main/java/com/minecraftonline/griefalert/inventories/InspectionInventoryStacks.java
@@ -26,9 +26,10 @@ package com.minecraftonline.griefalert.inventories;
 
 import com.google.common.collect.Lists;
 import org.spongepowered.api.data.key.Keys;
-import org.spongepowered.api.data.manipulator.mutable.DyeableData;
 import org.spongepowered.api.data.type.DyeColors;
 import org.spongepowered.api.item.ItemTypes;
+import org.spongepowered.api.item.enchantment.Enchantment;
+import org.spongepowered.api.item.enchantment.EnchantmentTypes;
 import org.spongepowered.api.item.inventory.ItemStack;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
@@ -39,7 +40,7 @@ public final class InspectionInventoryStacks {
   }
 
   public static final ItemStack INFO = ItemStack.builder()
-      .itemType(ItemTypes.GLASS)
+      .itemType(ItemTypes.BOOK)
       .add(Keys.DISPLAY_NAME, Text.of(TextColors.YELLOW, "Alert Info"))
       .add(Keys.ITEM_LORE, Lists.newArrayList(Text.of(TextColors.GRAY, "/ga info <index>")))
       .build();
@@ -62,9 +63,19 @@ public final class InspectionInventoryStacks {
       .build();
 
   public static final ItemStack QUERY = ItemStack.builder()
-      .itemType(ItemTypes.IRON_SHOVEL)
+      .itemType(ItemTypes.GOLDEN_SHOVEL)
       .add(Keys.DISPLAY_NAME, Text.of(TextColors.LIGHT_PURPLE, "Query Griefer"))
       .add(Keys.ITEM_LORE, Lists.newArrayList(Text.of(TextColors.GRAY, "/ga query -p <griefer>")))
+      .build();
+
+  public static final ItemStack TOOL = ItemStack.builder()
+      .itemType(ItemTypes.GLASS)
+      .add(Keys.DISPLAY_NAME, Text.of(TextColors.DARK_GREEN, "GriefAlert Tool"))
+      .add(Keys.ITEM_LORE, Lists.newArrayList(Text.of(TextColors.GRAY, "/ga tool <griefer>")))
+      .add(Keys.ITEM_ENCHANTMENTS, Lists.newArrayList(Enchantment.builder()
+          .type(EnchantmentTypes.MENDING)
+          .level(1).build()))
+      .quantity(1)
       .build();
 
   public static final ItemStack NONE = ItemStack.builder()